сделала правки по домашке и разбила на модули
